fix(cart): home and shop page add to cart button was not working
